Feat/  Reorganisation des extensions + room socket
