+++ app/celery.py
@@ -4,4 +4,4 @@ from app import create_app
 app = create_app(is_worker=True)
 app.app_context().push()
 
-from app import celery
+from app.extensions import celery
