+++ app/home/routes.py
@@ -1,7 +1,6 @@
 from app.home import bp
 from flask_login import login_required, current_user
 from flask import render_template, jsonify, request, current_app
-from app.extensions import db
 from app.models.favoris import FAVORIS
 from app.models.fichier import FICHIER
 from app.models.lien import LIEN
@@ -11,8 +10,9 @@ from datetime import datetime
 from app.models.dossier import DOSSIER
 from app.utils import Whoosh, check_notitications
 from app.decorators import active_required
-from app import socketio
+from app.extensions import socketio, db
 from fasteners import InterProcessLock
+from flask_socketio import join_room
 
 
 @bp.route("/", methods=["GET", "POST"])
@@ -56,6 +56,7 @@ def home():
         query=query,
         form=form,
         title="Accueil" if not query else query,
+        user_id=current_user.id_Utilisateur,
     )
 
 
@@ -227,6 +228,18 @@ def is_accessible(folder):
     """
     return any(current_user.id_Role == role.id_Role for role in folder.ROLE)
 
+@socketio.on("join", namespace="/home")
+def on_join(data):
+    """
+    Join a room.
+
+    Args:
+        data (dict): A dictionary containing the room information.
+
+    Returns:
+        None
+    """
+    join_room(data["room"])
 
 @socketio.on("search_files", namespace="/home")
 def search_files(data):
@@ -251,4 +264,5 @@ def search_files(data):
             "folderId": data.get("folderId"),
         },
         namespace="/home",
+        room=f"user_{current_user.id_Utilisateur}",
     )
