+++ app/mail.py
@@ -1,5 +1,6 @@
 import smtplib
 import os
+from config import Config
 from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
 
@@ -10,18 +11,16 @@ def send_email(subject, sender, recipients, html_body):
     msg["From"] = sender
     msg["To"] = recipients
     msg.attach(MIMEText(html_body, "html"))
-    mail_server = smtplib.SMTP(
-        os.environ.get("MAIL_SERVER"), os.environ.get("MAIL_PORT")
-    )
+    mail_server = smtplib.SMTP(Config.MAIL_SERVER, Config.MAIL_PORT)
     mail_server.starttls()
-    mail_server.login(os.environ.get("MAIL_USERNAME"), os.environ.get("MAIL_PASSWORD"))
+    mail_server.login(Config.MAIL_USERNAME, Config.MAIL_PASSWORD)
     mail_server.sendmail(sender, recipients, msg.as_string())
     mail_server.quit()
 
 
 def send_registration_request_email(email):
     subject = "Demande d'inscription en attente de validation"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = "<p>Bonjour, votre demande d'inscription est en attente de validation. Vous recevrez un email de confirmation dès que votre compte sera activé.</p>"
     send_email(subject, sender, recipients, html_body)
@@ -29,7 +28,7 @@ def send_registration_request_email(email):
 
 def send_registration_confirmation_email(email, password):
     subject = "Inscription validée"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = f"<p>Bonjour, votre demande d'inscription a été validée. Vous pouvez désormais vous connecter à la plateforme.</p><p>Votre mot de passe est: {password}</p>"
     send_email(subject, sender, recipients, html_body)
@@ -37,7 +36,7 @@ def send_registration_confirmation_email(email, password):
 
 def send_registration_rejection_email(email):
     subject = "Demande d'inscription refusée"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = "<p>Bonjour, votre demande d'inscription a été refusée.</p>"
     send_email(subject, sender, recipients, html_body)
@@ -45,7 +44,7 @@ def send_registration_rejection_email(email):
 
 def send_reactivation_confirmation_email(email):
     subject = "Réactivation de votre compte"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = "<p>Bonjour, votre compte a été réactivé. Vous pouvez désormais vous connecter à la plateforme.</p>"
     send_email(subject, sender, recipients, html_body)
@@ -53,7 +52,7 @@ def send_reactivation_confirmation_email(email):
 
 def send_reactivation_rejection_email(email):
     subject = "Réactivation de votre compte refusée"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = f"<p>Bonjour, la réactivation de votre compte a été refusée.</p>"
     send_email(subject, sender, recipients, html_body)
@@ -69,7 +68,7 @@ def send_forgotten_password_email(email, uuid):
     Returns: None
     """
     subject = "Réinitialisation de votre mot de passe"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = f"<p>Bonjour, une demande de réinitialisation de mot de passe a été faite sur votre compte. \n Pour la valider, cliquez sur le lien suivant :</p> <a href='http://127.0.0.1:5000/connexion/reinitialisation/{uuid}'>Réinitialisation du mot de passe</a>"
     send_email(subject, sender, recipients, html_body)
@@ -85,7 +84,7 @@ def send_deactivation_email(email):
         None
     """
     subject = "Désactivation de votre compte"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = "<p>Bonjour, votre compte a été désactivé. Pour plus d'informations, veuillez contacter le support technique.</p> <p>Vous pouvez demander la réactivation de votre compte depuis la page de connexion sur le lien suivant :</p> <a href='http://127.0.0.1:5000/connexion/'>Réactivation du compte</a>"
     send_email(subject, sender, recipients, html_body)
@@ -101,7 +100,7 @@ def send_delete_email(email):
         None
     """
     subject = "Suppression de votre compte"
-    sender = f'Hadès <{os.environ.get("MAIL_USERNAME")}>'
+    sender = f'Hadès <{Config.MAIL_USERNAME}>'
     recipients = email
     html_body = "<p>Bonjour, votre compte a été supprimé. Pour plus d'informations, veuillez contacter le support technique.</p>"
     send_email(subject, sender, recipients, html_body)
