+++ app/models/notification.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Integer, DateTime, Index, ForeignKeyConstraint, UUID
 from sqlalchemy.orm import relationship, mapped_column, Mapped
 from typing import Optional
-from app import db
+from app.extensions import db
 
 
 class NOTIFICATION(db.Model):
