+++ app/templates/desktop/index.html
@@ -23,8 +23,10 @@
 <script type="module" src="{{ url_for('static', filename='js/desktop.js') }}"></script>
 <script type="module" src="{{ url_for('static', filename='js/preview.js') }}" id="preview-script"></script>
 
-<!-- csrf token -->
+<!-- meta -->
 <meta name="csrf-token" content="{{ csrf_token() }}">
+<meta name="current-user" content="{{ user_id }}">
+
 {% endblock %}
 
 {% block content %}
