+++ app/templates/home/index.html
@@ -26,8 +26,9 @@
 <script type="module" src="{{ url_for('static', filename='js/home.js') }}"></script>
 <script type="module" src="{{ url_for('static', filename='js/preview.js') }}"></script>
 
-<!-- csrf-token -->
+<!-- meta -->
 <meta name="csrf-token" content="{{ csrf_token() }}">
+<meta name="current-user" content="{{ user_id }}">
 
 {% endblock %}
