+++ requirements.txt
@@ -24,3 +24,4 @@ celery==5.3.6
 fasteners==0.19
 flask_redis==0.4.0
 fr_hades_model @ https://github.com/Nathan-Etave/Hades/releases/download/fr_hades_model-1.3.0/fr_hades_model-1.3.0-py3-none-any.whl
+gunicorn==22.0.0
